Проект готов к проверке.

+++ src/ru/avalon/java/dev/j10/labs/shapes/Point.java
@@ -31,7 +31,7 @@ public interface Point {
      *
      * @see <a href="https://ru.wikipedia.org/wiki/%D0%90%D0%B1%D1%81%D1%86%D0%B8%D1%81%D1%81%D0%B0">Абсцисса</a>
      */
-    float getX();
+    int getX();
 
     /**
      * Возвращает ординату точки.
@@ -43,5 +43,5 @@ public interface Point {
      *
      * @see <a href="https://ru.wikipedia.org/wiki/%D0%9E%D1%80%D0%B4%D0%B8%D0%BD%D0%B0%D1%82%D0%B0">Ордината</a>
      */
-    float getY();
+    int getY();
 }
